// src/utils/enums/sidebar-message-type.enum.ts
export enum SidebarMessageType {
  SIGN_IN_WITH_PAT = 'signInWithPat',
  SIGN_OUT = 'signOut',
  REFRESH = 'refresh',
  FORCE_RELOAD = 'forceReload',
  SEARCH = 'search',
  SET_SERVER_URL = 'setServerUrl',
  OPEN_ISSUE = 'openIssue',
  OPEN_IN_BROWSER = 'openInBrowser',
  TOGGLE_MY_ISSUES = 'toggleMyIssues',
  APPLY_SAVED_FILTER = 'applySavedFilter',
  SAVE_FILTER = 'saveFilter',
}